fix utils, asc debut

+++ simulator/scenelib/button.ml
@@ -2,5 +2,5 @@ type t = {
   name: string;
   f: unit->bool;
   pos: Vector.t;
-  size: float;
+  size: int;
 }
